fix(hydration): extract Dialog from .map() rows (#290)
Radix Dialog generates internal IDs that can drift between server and
client render when the Dialog is rendered conditionally inside a list
.map() — Tobias hit this on /admin/termin where EditTerminForm's Pencil
icon flickered to a calendar icon and aria-controls IDs diverged.

Two patterns applied, following Tobias's AttendenceDialogUI prototype:

1. Split: EditTerminForm now has a thin wrapper that owns Dialog + useId
   and renders EditTerminFormUI (the actual form). The new
   TerminScheduleDialogUI does the same for the schedule Dialog that was
   inlined in TerminItem (the riskiest spot — parent is a server
   component in a map).

2. Patch: every other Dialog component that's already in its own client
   file but rendered inside a .map() gets useId() + id={id} on
   DialogContent for stable identity across re-renders.

Sites covered:
- admin/termin/{TerminItem,EditTerminForm,EditCourseToSchemaForm}
- admin/courses/forms/EditCourseForm
- admin/events/EditEventBtn
- admin/lectures/EditLesson
- admin/products/{AddCoursesToProductForm,EditProductForm}
- admin/students/{StudentUserEditDialog, StudentTableClient.CountDialogButton}
- admin/users/EditUserDialog
- public/courses/CourseInfoDialog

Skipped (Dialog rendered once outside any map): omoss/* lists,
legal/LegalPageList, students/MailDialog, all Add*Form dialogs.

Closes #290.
