+++ src/app/(admin)/admin/products/components/AddCoursesToProductForm.tsx
@@ -4,7 +4,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { DialogDescription } from "@radix-ui/react-dialog";
 import { Info, Pencil, Plus, X } from "lucide-react";
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
+import { useEffect, useId, useState } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 import type z from "zod";
@@ -69,6 +69,7 @@ export default function AddCoursesToProductForm({
   count,
   initialLang = "sv",
 }: Props) {
+  const id = useId();
   const form = useForm<CourseFormInput, unknown, CourseFormOutput>({
     resolver: zodResolver(formSchema),
     defaultValues: {
@@ -124,7 +125,7 @@ export default function AddCoursesToProductForm({
         </Button>
       </DialogTrigger>
 
-      <DialogContent className="max-h-[90dvh] overflow-auto">
+      <DialogContent id={id} className="max-h-[90dvh] overflow-auto">
         <DialogHeader>
           <DialogTitle>Lägg till kurser i produkt</DialogTitle>
           <DialogDescription>
