+++ src/app/(public)/courses/components/CourseInfoDialog.tsx
@@ -2,6 +2,7 @@
 
 import { ArrowUpRight } from "lucide-react";
 import Image from "next/image";
+import { useId } from "react";
 import {
   Dialog,
   DialogContent,
@@ -35,6 +36,7 @@ interface CourseInfoDialogProps {
 }
 
 export function CourseInfoDialog({ course }: CourseInfoDialogProps) {
+  const id = useId();
   const title = getCourseName(course);
 
   return (
@@ -45,7 +47,7 @@ export function CourseInfoDialog({ course }: CourseInfoDialogProps) {
           <ArrowUpRight className="w-3 h-3 shrink-0" />
         </span>
       </DialogTrigger>
-      <DialogContent className="max-h-[90dvh] overflow-auto">
+      <DialogContent id={id} className="max-h-[90dvh] overflow-auto">
         <DialogHeader>
           <DialogTitle>{title}</DialogTitle>
           <DialogDescription>{course.description}</DialogDescription>
